Complete the value-descriptor migration: registry guards, i128 caps, DeFi parity
Closes the remaining gaps of the typed value-descriptor migration:

- Criteria no longer carry per-tool match arms. The dead arms targeting
  an unregistered bridge tool are removed, and the soroban resource-fee
  criterion is removed entirely: its tool-name gate matched no
  registered tool, so it could never fire, and a guard that cannot fire
  misleads more than it protects. Policy files referencing it now fail
  to load with the unknown-criterion error; a future contract-invoking
  tool should get a descriptor-based resource criterion instead.
- Registry enumeration tests make the descriptor system self-guarding:
  every registered value-moving tool provably resolves to a non-empty
  value derivation (argument-derived tools through the shared derivation,
  handler-supplied tools through their shared leg builders), in both
  directions, so a forgotten tool is a loud test failure.
- Value caps compare and report in full i128: the deny-reason amount
  fields carry exact decimal strings, the loader parses caps beyond
  i64::MAX, and every narrowing clamp on a comparison path is gone. The
  persisted per-period window accumulator remains i64-width; a
  per-period rule refuses fail-closed rather than silently saturating
  when a call's resulting window total would exceed that bound.
- The per-domain value-leg builders live in their domain crates and are
  the single source for both surfaces: the MCP handlers import them, and
  the CLI trade, lend, and vault verbs now build the same legs from the
  same parsed values they submit, gating through the shared with-value
  evaluation helper; trustline gates through the shared argument-path
  builder. CLI DeFi calls are value-sized under a v1 policy, with
  refusal codes matching the MCP twins.
- The Blend request-type classification is exhaustive in its defining
  crate with no wildcard, so a future request variant is a compile error
  forcing an explicit debit classification.

Closes #22
Refs #20

+++ crates/stellar-agent-core/src/policy/v1/criteria/rate_limit.rs
@@ -131,18 +131,18 @@ impl Criterion for RateLimitCriterion {
         // state_key derivation ("rate_limit" bucket) so each inner in a bundle
         // contributes +1 call via the overlay.  On the single-tx path
         // (bundle = None) the overlay contributes 0.
-        let bundle_accumulated_calls: i64 = ctx
+        let bundle_accumulated_calls: i128 = ctx
             .bundle
             .map(|view| view.overlay.get(&state_key))
             .unwrap_or(0);
 
         // `.max(0)` is defence-in-depth: `BundleStateOverlay::accumulate` uses
-        // `i64::saturating_add` with positive addends (call counts), so the
+        // `i128::saturating_add` with positive addends (call counts), so the
         // overlay value should never be negative.  The clamp protects against
         // future criterion implementations that might accumulate negative amounts
         // via `accumulate_overlay`, preventing a negative overlay from
         // under-counting the window total and silently raising the effective cap.
-        // The subsequent saturating cast to u32 handles any `i64` value that
+        // The subsequent saturating cast to u32 handles any `i128` value that
         // would overflow u32.
         let calls_in_window = calls_in_window_recorded
             .saturating_add(u32::try_from(bundle_accumulated_calls.max(0)).unwrap_or(u32::MAX));
@@ -409,7 +409,7 @@ mod tests {
 
         // Overlay simulates 2 calls already approved in earlier bundle inners.
         let mut overlay = BundleStateOverlay::default();
-        // The overlay uses rate_limit key with i64 count.
+        // The overlay uses rate_limit key with i128 count.
         overlay.accumulate(rate_key.clone(), 2);
         let inners: Vec<crate::policy::v1::bundle::InnerOpDescriptor> = vec![];
         let view = BundleView {
